services/common/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools>=68", "wheel" ]

[project]
name = "kitty-common"
version = "0.1.0"
description = "Shared utilities for KITTY services"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "KITTY Team" },
]
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "alembic>=1.13",
  "anyio>=4.4",
  "httpx>=0.27",
  "paho-mqtt>=1.6",
  "passlib[bcrypt]>=1.7",
  "psycopg[binary]>=3.2",
  "pyasn1>=0.6",
  "pyasn1-modules>=0.3",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "redis>=5",
  "sqlalchemy>=2",
  "structlog>=24.1",
  "typing-extensions>=4.9",
  "zeroconf>=0.132",
]

optional-dependencies.dev = [
  "mypy>=1.11",
  "pytest>=8.1",
  "ruff>=0.6",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [ "src" ]

[tool.ruff]
line-length = 100

format.indent-style = "space"
format.quote-style = "double"
format.line-ending = "lf"
lint.select = [ "B", "E", "F", "I" ]
lint.ignore = [ "E203", "E266" ]

[tool.pytest.ini_options]
testpaths = [ "tests" ]
